requirements.txt
streamlit
requests
python-dotenv
pytz
google-generative-ai